sys-libs/hardened_malloc: add multilib
Add multilib compatibility to hardened_malloc. This doesn't actually add
multilib support for the package, but it provides a failsafe when
linking directly to hardened_malloc on multilib packages. A lot of
multilib packages will fail during the configure phase when the user has
specified linking hardened_malloc directly in LDFLAGS because a multilib
version of it doesn't exist. To workaround this, install a dummy library
for hardened_malloc on all non-native ABIs.